Fix on structural detector. Extract package refatoring.
