+++ src/main/oldWindowConstruction/testSmellPanel/ClassWithLackOfCohesionPanel.java
@@ -18,7 +18,7 @@ import it.unisa.testSmellDiffusion.utility.FolderToJavaProjectConverter;
 import main.refactor.IRefactor;
 import main.refactor.strategy.LackOfCohesionStrategy;
 import main.testSmellDetection.testSmellInfo.lackOfCohesion.LackOfCohesionInfo;
-import main.oldWindowConstruction.ContextualAnalysisFrame;
+import main.windowCommitConstruction.contextualAnalysisPanel.ContextualAnalysisFrame;
 
 import javax.swing.*;
 import javax.swing.border.Border;
