+++ src/main/windowCommitConstruction/contextualAnalysisPanel/ContextualAnalysisResultsFrame.java
@@ -1,4 +1,4 @@
-package main.oldWindowConstruction;
+package main.windowCommitConstruction.contextualAnalysisPanel;
 
 import org.repodriller.domain.Commit;
